save cart endpoint modifications and payment page integrated with backend

+++ app/controllers/datarecord.py
@@ -112,24 +112,32 @@ class DataRecord():
             for user in data:
                 if user["username"] == username:
                     return user
+            return None
         except FileNotFoundError:
             return None
     
 
-    def saveNewDates(self, username, newDates):
+    def saveUserCart(self, username, cart):
         try:
-
             with open("app/controllers/db/user_accounts.json", "r+") as arquivo_json:
                 data = json.load(arquivo_json)
+                
+                user_found = False
 
                 for user in data:
-                    if user["username"] == username:
-                        user.update(newDates)
+                    if user['username'] == username:
+                        user['cart'] = cart
+                        user_found = True
                         break
-                
+
+                if not user_found:
+                    print(f"Usuário '{username}' não encontrado.")
+                    return False
+
                 arquivo_json.seek(0)
-                json.dump(data, arquivo_json)
+                json.dump(data, arquivo_json, indent=4)
                 arquivo_json.truncate()
+                return True
 
         except FileNotFoundError:
             print("Arquivo não encontrado.")
@@ -138,6 +146,22 @@ class DataRecord():
         except Exception as e:
             print(f"Ocorreu um erro: {e}")
 
+    def getUserCart(self, username):
+        try:
+            user = self.getUserAccountDates(username)
+            print('entrou no getusertCart')
+            if user:
+                return user.get('cart', [])
+            else:
+                print(f"Usuário '{username}' não encontrado.")
+                return False
+
+        except FileNotFoundError:
+            print("Arquivo não encontrado.")
+        except json.JSONDecodeError:
+            print("Erro ao decodificar o JSON.")
+        except Exception as e:
+            print(f"Ocorreu um erro: {e}")
 
     '''def syncBags(userBag, tempBag):
         for tempItem in tempBag:
